Fix string/int comparison error in all verification controllers
- Convert timeout parameters from string to int in all verification controllers
- Convert other numeric parameters to appropriate types (int for timeouts/durations, float for thresholds)
- Fixes '>' not supported between instances of 'str' and 'int' error
- Applied to: adb.py, text.py, image.py, appium.py, video.py

+++ virtualpytest/src/controllers/verification/adb.py
@@ -511,7 +511,7 @@ class ADBVerificationController(VerificationControllerInterface):
                 }
             
             # Optional parameters with defaults
-            timeout = float(params.get('timeout', 0.0))
+            timeout = int(params.get('timeout', 0))
             
             print(f"[@controller:ADBVerification] Executing {command} with search term: '{search_term}'")
             print(f"[@controller:ADBVerification] Parameters: timeout={timeout}")
